separating to classes
