+++ lib/app/repositories/i_post_repository.dart
@@ -1,0 +1,14 @@
+import 'package:flutter_getx_template/app/data/models/post.dart';
+
+abstract class IPostRepository {
+  Future<List<Post>> getAll();
+
+  Future<Post> add({
+    required String text,
+    required String creationDate,
+  });
+
+  Future<Post> edit(Post post);
+
+  Future<bool> delete(int id);
+}
